refactor : api관련 타입 분리
