+++ src/hooks/query/event/useAppliedEventQuery.ts
@@ -1,5 +1,5 @@
 import getAppliedEvent from '@/apis/event/getAppliedEvent';
-import { GetAppliedEventRequest } from '@/types/event';
+import { GetAppliedEventRequest } from '@/types/api/getAppliedEvent';
 
 import { useQuery } from '@tanstack/react-query';
