+++ src/types/event.ts
@@ -28,32 +28,6 @@ interface Event {
   clubLogoImageUrl: string;
 }
 
-interface GetAppliedEventData {
-  id: string;
-  title: string;
-  clubName: string;
-  startDate: string;
-  location: string;
-  status: EventTagKey;
-  posterImageUrl: string;
-}
-
-interface GetAppliedEventResponse {
-  data: GetAppliedEventData[];
-  pageData: {
-    first: boolean;
-    last: boolean;
-    pageNumber: number;
-    size: number;
-    totalPages: number;
-    totalElements: number;
-  };
-}
-
-interface GetAppliedEventRequest {
-  pageNumber: number;
-}
-
 interface GetAllEventsRequest {
   pageNumber: number;
 }
@@ -108,8 +82,6 @@ type ProfileEventType = 'applied' | 'bookmark';
 export {
   GetAllEventsRequest,
   GetAllEventsResponse,
-  GetAppliedEventRequest,
-  GetAppliedEventResponse,
   getEventDetailResponse,
   FormOption,
   Event,
